Reorganize structure
