+++ configs/rand2char/glyffuser.py
@@ -1,12 +1,12 @@
 from dataclasses import dataclass
 
-from configs.base_config import TrainingConfigBase
+from configs.base_config import TrainConfigBase
 
 
 @dataclass
-class TrainingConfigChar2CharBi(TrainingConfigBase):
+class TrainConfig_R2C_Glyff(TrainConfigBase):
     """ Model Params """
-    task_name: str = "char2char_bi"             # Task name
+    task_name: str = "rand2char"                # Task name
     encoder_dim: int = 1024                     # Encoder dimensions
     
     """ Training Params (moved from base) """
