--- glyffuser/README.md
@@ -1,9 +1,0 @@
-# Glyffuser - text-to-image for Chinese glyphs
-
-See the blog post at https://yue-here.com/posts/glyffuser for further details
-
-Run the "Unihan glyph generator.ipynb" notebook to generate a local dataset.
-
-Run "glyffuser inference.ipynb" to generate characters using the trained model.
-
-"glyffuser conditional.ipynb" and "glyffuser unconditional.ipynb" contain the training code.
