+++ palette/char2char.json
