--- palette/dataset/__init__.py
@@ -1,1 +1,0 @@
-from palette.dataset.dataset import Char2CharDataset
