+++ palette/models/base_model.py
@@ -11,7 +11,7 @@ from torch.utils.data import DataLoader
 from palette.models.base_network import BaseNetwork
 from palette.utils.device_utils import set_device
 from palette.utils.logger import InfoLogger, MetricsLogger
-from configs.palette_config import TrainingConfigPalette
+from configs.char2char.palette import TrainConfig_C2C_Palette
 
 CustomResult = namedtuple("CustomResult", "name result")
 
@@ -19,7 +19,7 @@ CustomResult = namedtuple("CustomResult", "name result")
 class BaseModel:
     def __init__(
         self,
-        config: TrainingConfigPalette,
+        config: TrainConfig_C2C_Palette,
         phase_loader: DataLoader,
         val_loader: DataLoader,
         metrics: List[Callable],
